+++ src/ts/components/ads/AdCounterLabel.ts
@@ -3,10 +3,12 @@ import { UIInstanceManager } from '../../UIManager';
 import { LabelConfig, Label } from '../labels/Label';
 import { PlayerAPI } from 'bitmovin-player';
 import { StringUtils } from '../../utils/StringUtils';
+import { AdBreakTracker, AdBreakTrackerAdCountChangedArgs } from '../../utils/AdBreakTracker';
 
 export interface AdCounterLabelConfig extends LabelConfig {
   /**
-   * Message displayed during the ad indicating which ad out of how many in the current ad break is currently playing.
+   * Message displayed during the ad indicating which ad out of how many is currently playing. It takes all ad breaks
+   * with the same schedule time into account.
    * Supported placeholders: look at {@link StringUtils.replaceAdMessagePlaceholders}
    */
   adCountOutOfTotal?: LocalizableText;
@@ -19,15 +21,7 @@ export interface AdCounterLabelConfig extends LabelConfig {
  */
 export class AdCounterLabel extends Label<AdCounterLabelConfig> {
   private player?: PlayerAPI;
-  private updateLabelText = () => {
-    if (!this.player) {
-      return;
-    }
-
-    this.setText(
-      StringUtils.replaceAdMessagePlaceholders(i18n.performLocalization(this.config.adCountOutOfTotal), this.player),
-    );
-  };
+  private adBreakTracker?: AdBreakTracker;
 
   constructor(config: AdCounterLabelConfig = {}) {
     super(config);
@@ -46,20 +40,53 @@ export class AdCounterLabel extends Label<AdCounterLabelConfig> {
     super.configure(player, uimanager);
     this.player = player;
 
-    const clearText = () => {
-      this.setText('');
-    };
+    this.adBreakTracker = uimanager.getConfig().adBreakTracker;
+    this.adBreakTracker.onAdCountChanged.subscribe(this.adBreakTrackerAdCountChangedHandler);
 
-    player.on(player.exports.PlayerEvent.AdStarted, () => {
-      this.updateLabelText();
-    });
-    player.on(player.exports.PlayerEvent.AdBreakStarted, clearText);
-    player.on(player.exports.PlayerEvent.AdBreakFinished, clearText);
+    // An ad break may already be ongoing when configure is called, in this case the onAdCountChanged event was missed
+    // and the label is set here
+    if (this.adBreakTracker.currentAdIndex > 0 && this.adBreakTracker.totalNumberOfAds > 0) {
+      this.setAdCounterFromAdBreakTracker(this.adBreakTracker.currentAdIndex, this.adBreakTracker.totalNumberOfAds);
+    }
+  }
+
+  release(): void {
+    this.adBreakTracker?.onAdCountChanged.unsubscribe(this.adBreakTrackerAdCountChangedHandler);
+
+    this.adBreakTracker = undefined;
+    this.player = undefined;
+
+    super.release();
   }
 
   protected onLanguageChanged(): void {
-    if (this.player?.ads?.isLinearAdActive?.()) {
-      this.updateLabelText();
+    if (this.adBreakTracker?.currentAdIndex > 0 || this.player?.ads?.isLinearAdActive?.()) {
+      this.setAdCounterFromAdBreakTracker(this.adBreakTracker?.currentAdIndex, this.adBreakTracker?.totalNumberOfAds);
     }
+  }
+
+  private readonly adBreakTrackerAdCountChangedHandler = (
+    _: AdBreakTracker,
+    adBreakTrackerEvent: AdBreakTrackerAdCountChangedArgs,
+  ) => {
+    this.setAdCounterFromAdBreakTracker(adBreakTrackerEvent.currentAdIndex, adBreakTrackerEvent.totalNumberOfAds);
+  };
+
+  private setAdCounterFromAdBreakTracker(currentAdIndex?: number, totalNumberOfAds?: number) {
+    if (currentAdIndex === 0 && totalNumberOfAds === 0) {
+      // No ad break active and no subsequent ad breaks
+      this.setText('');
+      return;
+    }
+
+    this.setText(
+      StringUtils.replaceAdMessagePlaceholders(
+        i18n.performLocalization(this.config.adCountOutOfTotal),
+        this.player,
+        undefined,
+        currentAdIndex,
+        totalNumberOfAds,
+      ),
+    );
   }
 }